Refactor namespaces, Add Services & Models
