+++ test/cswm.Core.Test/WindowManagerTest.cs
@@ -1,5 +1,5 @@
-using System;
-using cswm.Core.Test.Management;
+using cswm.Core.Models;
+using cswm.Core.Test.Services;
 
 namespace cswm.Core.Test;
